Move signature sets types to utils
