+++ packages/beacon-state-transition/src/util/index.ts
@@ -2,21 +2,22 @@
  * @module chain/stateTransition/util
  */
 
+export * from "./activation";
+export * from "./attestation";
+export * from "./balance";
+export * from "./block";
+export * from "./blockRoot";
+export * from "./committee";
+export * from "./domain";
+export * from "./duties";
 export * from "./epoch";
 export * from "./fork";
-export * from "./domain";
-export * from "./blockRoot";
-export * from "./block";
-export * from "./validator";
-export * from "./seed";
-export * from "./signingRoot";
-export * from "./committee";
-export * from "./attestation";
-export * from "./proposer";
-export * from "./balance";
-export * from "./validatorStatus";
-export * from "./duties";
-export * from "./slot";
-export * from "./activation";
-export * from "./shuffle";
 export * from "./genesis";
+export * from "./proposer";
+export * from "./seed";
+export * from "./shuffle";
+export * from "./signatureSets";
+export * from "./signingRoot";
+export * from "./slot";
+export * from "./validator";
+export * from "./validatorStatus";
