+++ packages/beacon-state-transition/src/util/signatureSets.ts
@@ -1,5 +1,24 @@
-import bls from "@chainsafe/bls";
-import {ISignatureSet, SignatureSetType} from "./types";
+import bls, {PublicKey} from "@chainsafe/bls";
+import {BLSSignature, Root} from "@chainsafe/lodestar-types";
+
+export enum SignatureSetType {
+  single = "single",
+  aggregate = "aggregate",
+}
+
+export type ISignatureSet =
+  | {
+      type: SignatureSetType.single;
+      pubkey: PublicKey;
+      signingRoot: Root;
+      signature: BLSSignature;
+    }
+  | {
+      type: SignatureSetType.aggregate;
+      pubkeys: PublicKey[];
+      signingRoot: Root;
+      signature: BLSSignature;
+    };
 
 export function verifySignatureSet(signatureSet: ISignatureSet): boolean {
   const signature = bls.Signature.fromBytes(signatureSet.signature.valueOf() as Uint8Array);
